Fix Swift 5 and some C++ warnings (#270)

+++ src/Bitcoin/Transaction.h
@@ -46,7 +46,7 @@ struct Transaction {
     bool empty() const { return inputs.empty() && outputs.empty(); }
 
     /// Generates the signature pre-image.
-    std::vector<uint8_t> getPreImage(const Script& scriptCode, int index, uint32_t hashType,
+    std::vector<uint8_t> getPreImage(const Script& scriptCode, size_t index, uint32_t hashType,
                                      uint64_t amount) const;
     std::vector<uint8_t> getPrevoutHash() const;
     std::vector<uint8_t> getSequenceHash() const;
